Adds error handling to component initialisation
Improves the resilience of component initialisation by adding try-catch blocks.

This prevents errors during the initialisation of the NSW Design System components from crashing the application and logs warnings to the console instead.
Also, adds `search` as a dependency for the `useEffect` hook in the header component to ensure the search functionality is initialized when the `search` prop changes.